Further AI updates:
When AI has no line of fire for any of their units, they will spread out on the map while keeping a line of fire towards a position from where they would have a line of fire.
Fixed issue with melee-units walking towards units that had a grenade below but then not attacking them. Now the targets with grenades below are ignored.